requirements.txt
Keras==2.2.1
matplotlib==3.3.4
numpy==1.19.1
opencv_python==3.4.2.16
Pillow==9.0.1
tensorflow_gpu==1.6.0
